requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
crypto==1.4.1
dataclasses-json==0.5.7
ECPy==0.10.0
idna==3.4
marshmallow==3.19.0
marshmallow-enum==1.5.1
mypy-extensions==1.0.0
Naked==0.1.32
packaging==23.1
PyJWT==2.7.0
PyYAML==6.0
requests==2.31.0
shellescape==3.8.1
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==2.0.2
